requirements.txt
# Google Cloud SDK
google-cloud-storage>=2.10.0
functions-framework>=3.4.0
google-cloud-logging>=3.5.0

# Web requests and API interactions
requests>=2.31.0
urllib3>=2.0.0
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0
python-dateutil>=2.8.0

# Geographic data (for PLU)
geopandas>=0.14.0
owslib>=0.29.0  # For WFS API interactions